test/pyproject.toml
[tool.poetry]
name = "test-tool"
version = "1.0.0"
description = ""
authors = ["Diego Alloza González <diego.allozagonzalez@telefonica.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.10"
behave = "^1.2.6"
allure-behave = "^2.9.45"
requests = "^2.28.0"
toolium = "^2.6.0"
python-dateutil = "^2.8.2"
jsonschema = "^4.6.0"
mysql-connector-python = "^8.0.29"
jq = "^1.2.2"
deepdiff = "^5.8.1"
Faker = "^13.15.1"
Telethon = "^1.24.0"
asyncio = "^3.4.3"
PyHamcrest = "^2.0.3"
ansible-vault = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
